+++ Changelog.md
@@ -4,6 +4,7 @@
 * [#4](https://github.com/vizor-games/jac/issues/4) Add implicit `^top` and `^base` profiles
 * Allow reference local values inside string evaluations at top level keys
 * Allow reference profile name at config evaluation
+* Profile keys shadow object methods in string evaluations
 
 # Version 0.0.3